+++ packages/protos/protos_weebi/CHANGELOG.md
@@ -1,10 +1,19 @@
+## 1.0.7+4 - 2025/01/03
+
+- ticket.creationDate added
+
+## 1.0.7+3
+
+- added logo in boutique and in BoutiqueRequest and in BoutiqueResponse
+- breaking change on readBoutique
+
 ## 1.0.7+2
 
 - removed authenticateWithDevice
 - removed readUserPermissionsByToken
 
-
 ## 1.0.7+1
+
 - addMany for article, contact, ticket
 - handle photo for articles
 - lastTouchTimestampUTC for lighter sync